file-structure-cleanup
