+++ cogs/__init__.py
@@ -1,0 +1,1 @@
+import dota,music,pogoda,rep,roulette,voice
