+++ cogs/icon.py
