+++ cogs/music.py
