+++ cogs/speech.py
